+++ packages/expo-router/src/fork/native-stack/createNativeStackNavigator.tsx
@@ -16,7 +16,6 @@ import {
   type StackNavigationState,
   StackRouter,
   type StackRouterOptions,
-  type StaticConfig,
   type TypedNavigator,
   useNavigationBuilder,
 } from '../../react-navigation/native';
@@ -184,7 +183,6 @@ export function createNativeStackNavigator<
     };
     Navigator: typeof NativeStackNavigator;
   },
-  const Config extends StaticConfig<TypeBag> = StaticConfig<TypeBag>,
->(config?: Config): TypedNavigator<TypeBag, Config> {
-  return createNavigatorFactory(NativeStackNavigator)(config);
+>(): TypedNavigator<TypeBag> {
+  return createNavigatorFactory(NativeStackNavigator)();
 }