+++ packages/expo-router/src/react-navigation/bottom-tabs/navigators/createBottomTabNavigator.tsx
@@ -3,7 +3,6 @@ import {
   createNavigatorFactory,
   type NavigatorTypeBagBase,
   type ParamListBase,
-  type StaticConfig,
   type TabActionHelpers,
   type TabNavigationState,
   TabRouter,
@@ -70,7 +69,6 @@ export function createBottomTabNavigator<
     };
     Navigator: typeof BottomTabNavigator;
   },
-  const Config extends StaticConfig<TypeBag> = StaticConfig<TypeBag>,
->(config?: Config): TypedNavigator<TypeBag, Config> {
-  return createNavigatorFactory(BottomTabNavigator)(config);
+>(): TypedNavigator<TypeBag> {
+  return createNavigatorFactory(BottomTabNavigator)();
 }